[IMP] web: form: don't block unload on new, non-dirty record
Before this commit, creating a new record in form view or in the
kanban record quick create, not touching any field and closing the
tab would result in a confirmation dialog blocking the tab closure,
like we do when the user is editing a record, but that record isn't
valid, so we can't save it.

However, in the case of a record creation that isn't dirty, there's
no point preventing the user from closing the tab. This commit thus
improves that behavior.

Part-of: odoo/odoo#228632
Related: odoo/enterprise#95562
Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>